+++ open_ahrs/src/common.rs
@@ -39,7 +39,7 @@ use quaternion::quaternion::Quaternion;
 use linalg::vector::dot_product;
 
 #[cfg(feature = "std")]
-//pub fn generate_random_attitudes(number: u8, ts: f64) -> Result<Matrix, OpenAHRSError> {
+//pub fn generate_random_attitudes(number: u8, ts: f32) -> Result<Matrix, OpenAHRSError> {
 pub fn generate_random_attitudes(number: u8) -> Result<Matrix, OpenAHRSError> {
     //extern crate quaternion;
     extern crate rand;
@@ -57,19 +57,19 @@ pub fn generate_random_attitudes(number: u8) -> Result<Matrix, OpenAHRSError> {
     previous_quat.fill_identity()?;
 
     for n in 0..number {
-        let u1: f64 = rand::thread_rng().gen();  // Generates a random number between 0 and 1.
-        let u2: f64 = rand::thread_rng().gen();  // Generates a random number between 0 and 1.
-        let u3: f64 = rand::thread_rng().gen();  // Generates a random number between 0 and 1.
+        let u1: f32 = rand::thread_rng().gen();  // Generates a random number between 0 and 1.
+        let u2: f32 = rand::thread_rng().gen();  // Generates a random number between 0 and 1.
+        let u3: f32 = rand::thread_rng().gen();  // Generates a random number between 0 and 1.
 
-        let s1: f64 = sqrt(1.0 - u1);
-        let s2: f64 = sqrt(u1);
-        let t1: f64 = 2.0 * PI * u2;
-        let t2: f64 = 2.0 * PI * u3;
+        let s1: f32 = sqrt(1.0 - u1);
+        let s2: f32 = sqrt(u1);
+        let t1: f32 = 2.0 * PI * u2;
+        let t2: f32 = 2.0 * PI * u3;
 
-        let mut qw: f64 = s2 * cos(t2);
-        let mut qx: f64 = s1 * sin(t1);
-        let mut qy: f64 = s1 * cos(t1);
-        let mut qz: f64 = s2 * sin(t2);
+        let mut qw: f32 = s2 * cos(t2);
+        let mut qx: f32 = s1 * sin(t1);
+        let mut qy: f32 = s1 * cos(t1);
+        let mut qz: f32 = s2 * sin(t2);
 
         let mut actual_quat: Quaternion = Quaternion::new()?;
         actual_quat.fill(qw, qx, qy, qz)?;
@@ -89,18 +89,18 @@ pub fn generate_random_attitudes(number: u8) -> Result<Matrix, OpenAHRSError> {
 
 
         // Perform scalar product.
-        let scalar: f64 = dot_product(&actual_quat.get_vect()?, &previous_quat.get_vect()?)?;
-        let theta: f64 = acos(scalar);  // Angle between the axis of the current quaternion and that of the previous quaternion.
+        let scalar: f32 = dot_product(&actual_quat.get_vect()?, &previous_quat.get_vect()?)?;
+        let theta: f32 = acos(scalar);  // Angle between the axis of the current quaternion and that of the previous quaternion.
 
         /*
             If the angle between the two axes of rotation of the quaternions is too great, an intermediate quaternion must be determined by spherical interpolation,
             ensuring that theangle between the preceding quaternion and the interpolated quaternion does not exceed the maximum angle.
         */
 
-        let max_angle: f64 = PI / 32.0_f64;
+        let max_angle: f32 = PI / 32.0_f32;
 
         if fabs(theta) > max_angle {
-            let scale: f64 = max_angle / fabs(theta);
+            let scale: f32 = max_angle / fabs(theta);
             let mut interpolated_quat: Quaternion = Quaternion::new()?;
             interpolated_quat.slerp(&previous_quat, &actual_quat, scale)?;
             qw = interpolated_quat.get_qw()?;
